move HistoryViewer to helpers.py, refactor imports
