+++ sync_threads/commands.py
@@ -1,5 +1,4 @@
 import os
-from typing import Sequence
 
 
 CommandNotFoundError = KeyError  # TODO: add implementation
@@ -20,20 +19,6 @@ class Commands:
             return command_func(*args)
         except KeyError:
             raise CommandNotFoundError
-
-
-class HistoryViewer:
-    def __init__(self, history: Sequence[str]):
-        self._history = history
-
-    @property
-    def as_strings(self):
-        return '\n'.join(
-            map(
-                lambda ind_command: f'{ind_command[0]:>5}  {ind_command[1]}',
-                enumerate(self._history, 1)
-                )
-            )
 
 
 @Commands.register
